+++ hbase-server/src/test/java/org/apache/hadoop/hbase/regionserver/TestCompactingMemStore.java
@@ -61,15 +61,15 @@ import org.junit.experimental.categories.Category;
 public class TestCompactingMemStore extends TestDefaultMemStore {
 
   private static final Log LOG = LogFactory.getLog(TestCompactingMemStore.class);
-  private static MemStoreChunkPool chunkPool;
-  private HRegion region;
-  private RegionServicesForStores regionServicesForStores;
-  private HStore store;
+  protected static MemStoreChunkPool chunkPool;
+  protected HRegion region;
+  protected RegionServicesForStores regionServicesForStores;
+  protected HStore store;
 
   //////////////////////////////////////////////////////////////////////////////
   // Helpers
   //////////////////////////////////////////////////////////////////////////////
-  private static byte[] makeQualifier(final int i1, final int i2) {
+  protected static byte[] makeQualifier(final int i1, final int i2) {
     return Bytes.toBytes(Integer.toString(i1) + ";" +
         Integer.toString(i2));
   }
@@ -81,6 +81,12 @@ public class TestCompactingMemStore extends TestDefaultMemStore {
 
   @Override
   public void setUp() throws Exception {
+    compactingSetUp();
+    this.memstore = new CompactingMemStore(HBaseConfiguration.create(), CellComparator.COMPARATOR,
+        store, regionServicesForStores);
+  }
+
+  protected void compactingSetUp() throws Exception {
     super.internalSetUp();
     Configuration conf = new Configuration();
     conf.setBoolean(SegmentFactory.USEMSLAB_KEY, true);
@@ -91,12 +97,10 @@ public class TestCompactingMemStore extends TestDefaultMemStore {
     this.region = hbaseUtility.createTestRegion("foobar", hcd);
     this.regionServicesForStores = region.getRegionServicesForStores();
     this.store = new HStore(region, hcd, conf);
-    this.memstore = new CompactingMemStore(HBaseConfiguration.create(), CellComparator.COMPARATOR,
-        store, regionServicesForStores);
+
     chunkPool = MemStoreChunkPool.getPool(conf);
     assertTrue(chunkPool != null);
   }
-
 
   /**
    * A simple test which verifies the 3 possible states when scanning across snapshot.
@@ -609,6 +613,11 @@ public class TestCompactingMemStore extends TestDefaultMemStore {
     while (((CompactingMemStore)memstore).isMemStoreFlushingInMemory()) {
       Threads.sleep(1000);
     }
+    int counter = 0;
+    for ( Segment s : memstore.getListOfSegments()) {
+      counter += s.getCellsCount();
+    }
+    assertEquals(3, counter);
     assertEquals(0, memstore.getSnapshot().getCellsCount());
     assertEquals(528, regionServicesForStores.getGlobalMemstoreTotalSize());
